+++ src/components/coursePageComponents/BadgeRatedEnrolled.css
