+++ src/components/coursePageComponents/BadgeRatedEnrolled.js
