+++ src/components/coursePageComponents/ContentAccordion.css
