+++ src/components/coursePageComponents/ContentAccordion.js
