+++ src/components/coursePageComponents/ContentItem.js
