+++ src/components/coursePageComponents/CourseContent.css
