+++ src/components/coursePageComponents/CourseContent.js
