+++ src/components/coursePageComponents/CourseFooter.css
