+++ src/components/coursePageComponents/CourseFooter.js
