+++ src/components/coursePageComponents/CourseHeader.js
