+++ src/components/coursePageComponents/ExpandableHTMLContainer.css
