+++ src/components/coursePageComponents/SideMenu.js
