+++ src/components/coursePageComponents/StickyInfoBar.css
