+++ src/components/coursePageComponents/StickyInfoBar.js
