+++ src/components/homePageComponents/CategoriesSection.css
