+++ src/components/homePageComponents/CoursesSlider.js
